info: fix creating StatePath from a List of path-elements Strings
The StatePath has a couple of static methods that provide convenient ways of creating StatePath objects.  One 
parses a string, breaking it into path-elements at each dot.  The second takes a List<String> of path-elements 
and creates the corresponding StatePath.  This second form is only used when processing 
InfoGetSerialisedDataMessage objects, which are requests for the current dCache status serialised into XML.

The StatePath class' equals method relies that all path-elements are intern()ed.  For all "normal" cases, this 
is true, but when the StatePath is built from a List<String> of path-elements, this is not true.  In ealier 
version of the code, this didn't matter as the equals() method is not used; however, with the new 
StateGuide-based VISITOR pattern, the StatePath.equals is used.  The result of this bug is that requesting XML 
data limited to a specific subtree would provide no data (beyond the top-level <dCache/>).

This patch fixes the issue: StatePath objects created by List<String> method will work correctly.

Patch is for trunk

RB: http://rb.dcache.org/r/1669/
Acked-by: Jan