gPlazma: Move gPlazma to slf4j.
This patch moves the core gPlazma code from log4j to slf4j. Part of this
is to remove the code for modifying the log level as this should be
orthogonal to the code that does the logging (it should be done through
the container, JMX, config files, etc).

Target: trunk
Acked-by: Paul Millar <paul.millar@desy.de>
Acked-by: Timur Perelmutov <timur@fnal.gov>
Patch: http://rb.dcache.org/r/1555/