Dcap GSI detection is inconsistent in header and library detection.
Due to patch, http://rb.dcache.org/r/1646/ not detecting one to the globus libraries for VDT it was observed that the "make" target built the gsi tunnel but with the wrong linking and the "make rpm " target failed as the gsi tunnel was not packaged in the RPM.

This was because the decision to build the gsi tunnel was based upon the globus header files being present and not requiring the libraries to be present also. While the packaging of the RPM for the gsi tunnel depended upon the globus header files being present and requiring the libraries to also contain the appropriate functions.

This patch resolves this. I use the auto tools convention or "$want_opt" = yes as I intend to move over to this standard in time.

Acked-by: Jan
http://rb.dcache.org/r/1673/