dcap: explicitly use the globus version of open ssl
This is a minimal changes fix, so the behaviour is the same as the last dynamic lib version of dcap, this explicitly uses the globus version of open ssl although a different approach may be needed with EPEL packaged globus testing is still required.

Probably this can be done better but it makes gsi work properly once this is done we can focus on getting the correct methods detected which we will work upon in the patch, http://rb.dcache.org/r/1646/

Consider this an essential bug fix for gsidcap release for all versions of globus packaged in the globus way.
Signed-off by: Paul
http://rb.dcache.org/r/1686/