info: update MalleableStateTransition to support additional testing
The MalleableStateTransition provides a framework for testing what happens when dCache's state is updated.

This patch adds updateTransitionChangingMetricToBranch() method.  This simulates a transition when an existing metric is updated 
so it becomes a branch.  The converse (a branch becoming a metric) is supported by the existing updateTransitionForNewMetric() 
method.  Tests for this new feature are added to the MalleableStateTransitionTests.

The patch also includes a bug-fix for updateTransitionForNewMetric() method where it doesn't work (NPE) if the metric is to be 
added to the root branch.  A test for this has been added to MalleableStateTransitionTests.

Patch is for trunk.

RB: http://rb.dcache.org/r/1677/
Acked-by: Gerd