info: add PostTransitionStateExhibitor
The PostTransitionStateExhibitor wraps a StateTransition and allows a visitor to visit what 
dCache's state will be like after the StateTransition has been applied.

The ability to visit the future dCache state is currently supplied by the StateComposite object 
directly, which complicates that class and how the VISITOR pattern is used.  The introduction of 
PostTransitionStateExhibitor class disentangles that, decoupling the two activities and allowing 
the StateComposite object to be simpler.  It also works towards allowing atomic updates of dCache 
state.

This patch does not result in the PostTransitionStateExhibitor being used---that's coming next.  
Instead, it adds the class and a fairly extensive set of unit-tests that achieve 100% coverage.

Patch is for trunk; the unit tests require RB #1677.

Target: trunk
RB: http://rb.dcache.org/r/1678
Acked-by: Gerd