dcap: silence object-to-fn-ptr warnings when compiling code
When compiling dcap, the GCC compiler emits warnings like:

    system_io.c:231: warning: ISO C forbids conversion of object pointer to function pointer type

This is because we are converting an object pointer (returned by dlsym() function), which we want to use as 
a function pointer.  This is arguably a bug in dlsym; however, we need a work-around in order to silence 
these warnings.

The workaround is to declare a union that does the conversion.  This is guaranteed (ISO/IEC 9899:1990 
"C89"/"C90") to match exactly.

The same problem exists in at least one other file.  Future patchs should address further occurrences of 
this problem.

Target: trunk
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/1680/