dcap: make door more robust
The dcap client library communicates with the dcap door by communicating a series of commands with an 
associated target.  The target is either a PNFS-ID or it is a URI.  Valid URIs have a particular form: 
<scheme>://<authority><path>, where <scheme> must end "dcap" (e.g., "dcap", "gsidcap"), <authority> 
must be non-empty (either <host> or <host>:<port>) and <path> must begin with '/' and also be 
non-empty.

Due a bug in the dcap client, it is possible to send a malformed requests to the dcap door.  One 
particular bug in dcap is triggered if the user specifies a URI like:

    pnfs://dcap-door/path/to/file

Note that the scheme is "pnfs" and not "dcap" (or "gsidcap").  When the dcap client receives such a 
URI, it sends "path/to/file" as the URI to the DCAP door.  Although this path is invalid, the dcap 
door will accept it, resulting in potential confusion, especially if the file is to be stored on tape.

This patch makes the dcap door more robust by rejecting badly formed URIs.

Patch is for trunk, 1.9.7 and 1.9.6.  I anticipate committing the patch to 1.9.5 after discussing this 
with sites.

Patch: http://rb.dcache.org/r/1649/
Acked-by: Timur Perelmutov <timur@fnal.gov>
Target: trunk
Target: 1.9.7
Target: 1.9.6
Target: 1.9.5