namespace: do not require OSMTemplate and sGroup in a directory
on a default installation we do not need a OSMTemplate and
sGroup tags to be present. Nevertheless dCache can't work with
such directories and forces admins to create then with
dummy values.

This patch drops this requirement and allows dCache to
operate on 'clean' namespace.

Target: Trunk
Acked-by: Gerd
Patch: http://rb.dcache.org/r/1689/