migration: fix storage-info check
The GenericStorageInfo equals method is sensitive to the contents of the StorageInfo "key" hash.

This feature is a PNFS-ism: it's a by-product of parsing a file's level-2 metadata.  Chimera
doesn't support this as it doesn't parse or update a file's level-2 metadata; however, simply
removing support for the file-metadata is controversial because we expose the flags as a parameter
supplied to the HSM script.

Currently, the "storage info" migration check fails because two storage elements (one from PNFS the
other from Chimera) are not equal().

My own point-of-view is that the hash-flags are a PNFS-ism and should be removed from the HSM
script.  An alternative method of providing the same information should be made available.
However, this could only be done on Trunk.

So, for the supported branches, we include this "work-around" to allow the test to succeed.

Patch is for all supported branches and trunk.

Target: 1.9.5
Target: 1.9.6
Target: 1.9.7
Target: Trunk
RB: http://rb.dcache.org/r/1705/
Acked-by: Tigran