info: add unit-tests for XML output
The XmlSerialiser (part of the info service) produces XML representation of dCache's current state.

This patch adds an initial set of tests to check that this works as expected.

The tests use Relax-NG to validate that the XML is as expected.  Rather than using a generic Relax-NG 
schema, each test has it's own schema that is much more strict on what the XML should look like; 
however, the validation does not check irrelevant aspects (like white-space, order of the attributes in 
an element, etc) so provides a much better test framework than a simple string comparator.

The Relax-NG compact syntax is used to record the schema to validate against.  This is much more terse 
than the equivalent in Relax-NG XML-format; and very much more compact than XML Schema.

The patch also introduces the jing jar file (license: BSD).  This is a Relax-NG parser/validator from 
the jing project:

    http://www.thaiopensource.com/relaxng/jing.html

The jar file may be run in a stand-along fashion (which may be useful for functional tests).  The unit 
tests use the Jing-specific API.

The patch adds this jar file to the JUnit class-path, but does not include it when building dCache (or 
other components) or package it in the RPM, tgz, etc.

Target: trunk
RB: http://rb.dcache.org/r/1707/
Acked-by: Jan