config: add support for forbidden property annotation
The current configuration system in trunk has only a single annotation to describe a property that is no 
longer supported in dCache: "(obsolete)".  The current behaviour of dCache is that, if a site has 
configuration that attempts to control dCache's behaviour by assigning a value to an obsolete property 
then dCache refuses to start.

This behaviour is arguably too harsh.  Some no-longer-supported properties are not dangerous: dCache will 
"Always Do the Right Thing" without requiring the information formally provided by the property.  An 
example of this is the ftpBase property; as the ftp door no longer requires the name-space to be mounted, 
dCache no longer relies on this property and will always work correctly.  Failing to start dCache in this 
case is needlessly disruptive.

The semantics of properties marked "(obsolete)" and "(forbidden)" are:

 obsolete:  indicates that a property is no longer supported and that dCache will
            always behaves correctly without supporting this property.  It is safe
            for dCache to allow attempt to assign a value to this property.

 forbidden: indicates that a property is no longer supported and dCache does not
            always behave correctly without further configuration or that support
            for some feature has been removed.  It would be unsafe for dCache to
            start with configuration that attempts to assign a value to this
            property.

If a dCache instance's configuration assigns a value to an obsolete property then a warning is emitted 
and dCache continues to load.  This is similar to the behaviour with deprecated properties.

If a dCache instance's configuration assigns a value to a forbidden property then dCache will refuse to 
start.

Patch is for trunk.

Patch: http://rb.dcache.org/r/1700
Acked-by: Gerd
Acked-by: Jan
Acked-by: Tigran
Target: trunk