config: add ReplaceableProperties unit-tests
We currently have no unit-tests that demonstrate whether ReplaceableProperties work as expected.

This patch introduces a set of tests that achieve ~78% coverage.

The noticeable omissions are load(File) and loadFromXML(String).  The former is currently a wrapper around 
load(Reader), which is tested, and the latter isn't used currently.

Patch is for trunk.

Patch: http://rb.dcache.org/r/1713/
Acked-by: Tigran
Target: Trunk