info: switch to using PostTransitionStateExhibitor
One core concept of the info service is that all the state is stored in a family of classes 
(those that implement StateComponent interface) and that other components are, as much as is 
possible, stateless.  Discovering state is achieved via the VISITOR pattern: a object from a 
class that implements the StateVisitor interface may acquire information by being presented 
to an object from a class that implements the StateExhibitor interface.

When dCache state is updated, the components that maintain derived metrics (the Secondary 
Information Providers: SIPs) may need to know information from both "before" and "after" the 
transition to figure out which metrics need to be updated.  Since these SIPs are stateless, 
they must acquire this information by visiting (via the VISITOR pattern) the current- and 
future- dCache state.

Currently, support for visiting the future (or "post transition") state is achieved 
explicitly within the classes that implement the StateComponent interface (StateComposite, 
StateValue, ...).  This is undesirable since it complicates the StateComponent classes and 
couples unrelated activities: storing state information and supporting visiting the state 
post-transition.

A recent patch introduced the PostTransitionStateExhibitor class, but in isolation: the 
class is not used anywhere.  The PostTransitionStateExhibitor class implements the 
StateExhibitor interface, allowing a StateVisitor to "visit" what will be the state of 
dCache after a transition: the future state.

The patch under review simply adjusts the code-base so it uses the 
PostTransitionStateExhibitor class.  At the heart of the patch is the change to the 
StateExhibitor interface: the method

  void visitState( StateVisitor visitor, StateTransition transition);

is removed as support for visiting the "future" dCache state is achieved by creating a 
PostTransitionStateExhibitor object.

As a knock-on effect, the StateWatcher interface (which the SIPs implement) has changed; the 
existing trigger method is replaced with the following:

    void trigger(StateUpdate update, StateExhibitor currentState, StateExhibitor 
futureState);

This means that SIPs are provided with the current- and future- state as StateExhibitors 
that may be used against any visitor.

The state visitors typically have static, utility methods for creating a visitor object, 
visiting the state and returning the collected information (typically a Map or a List).  
Prior to this patch, there would be two such static methods: one for the current state and 
one for the future state.  With this patch, discovering information about the future state 
is achieved in the same fashion as the current state, so the extra set of static methods are 
no longer needed.

A further knock-on effect is that classes that implement StateWatcher (such as all the SIPs) 
no longer need to be given a StateExhibitor in their constructor.  The patch leaves the 
existing constructors as-is, even though the StateExhibitor passed as a parameter is 
ignored.  This is deliberate ploy to reduce the size of the patch somewhat: a future patch 
will switch to using the default constructor.

Patch is for trunk.

Acked-by: Tigran
Patch: http://rb.dcache.org/r/1718/
Target: trunk