info: remove unnecessary constructors
A previous patch updated the info-service so it uses the PostTransitionStateExhibitor.

This patch builds on the previous patch by removing the StateExhibitor constructor for the 
SIPs.  With the use of PostTransitionStateExhibitor, it is no longer necessary to build a 
StateWatcher with a StateExhibitor as an argument.

Patch is for trunk.

Acked-by: Tigran
Patch: http://rb.dcache.org/r/1720/
Target: trunk