WebDAV: Make use of RedirectedTransfer
RedirectedTransfer allows the client to wait for a redirect message from
the pool. The class was introduced when the xrootd door was updated to
use the Transfer class, however the WebDAV door has not made use of
RedirectedTransfer until now.

The patch allows us to drop the BlocingQueueMap class and will make it
simpler to implement proxy mode for reads.

Target: trunk
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/1716/