xrootd: Make authorisation more flexible
As xrootd is unauthenticated, we do not have a Subject for the client
making the request. Until recently the xrootd door implemented its own
permission checking on write. The logic in this check was however
inconsistent with the rest of dCache.

This patch makes the authentication of the xrootd door more flexible:

* Allows both read and write paths to be restricted (previously only
  the write path could be restricted)

* Allows requests to be authenticated as any user (including nobody)


Notice that this authentication is independent of the token based
authentication performed for xrootd: That check only validates the
authentication information from the catalog. It does not actually
perform authentication within dCache wrt the files being accessed.


FOR RELEASE NOTES:

  This patch changes the interpretation of xrootdAllowedPaths. The paths
  now apply to both reading and writing. Two new properties have been
  added to allow reading and writing to be configured separately.

  The interpretation of the empty string for xrootdAllowedPaths has
  changed: An empty string used to allow writing to all paths. Now it
  disallows access to any path. To allow access to all paths the property
  has to be set to / (which is the default).

  The permission check on write has changed. In previous releases the parent
  directory had to have write and execute permissions set for the owner
  of the directory. With this patch both reads and writes are subject to
  authorisation checks for the user specified by the xrootdUser property.
  The default is nobody, meaning that only world readable/writeable files
  can be read/written.

Target: trunk
Ticket: #5538
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/1706/