info: remove TestStateExhibitor's now redundant support for cloning
The TestStateExhibitor provides a convenient, encapsulated and working StateExhibitor where new metrics can 
be added very simply.  This allows unit-tests to construct arbitrary dCache states without much fuss.

The current (in Trunk) version of this class includes support for cloning the tree structure.  This was 
needed to support visiting the state with a StateTransition; i.e., visiting the future state after a 
pending set of metric-additions, -removals and updates has taken place.

With the introduction of PostTransitionStateExhibitor, this functionality has been removed from 
StateExhibitor interface and the TestStateExhibitor no longer needs to clone itself.  However, in an effort 
to keep previous patches small, no effort was made to remove the cloning within TestStateExhibitor.

This patch removes the support for cloning the TestStateExhibitor.  It also removes the unit-tests that 
previously tested that this cloning process worked correctly.

Target: trunk
Patch: http://rb.dcache.org/r/1721/
Acked-by: Gerd