defaults: update header to use stronger language not to edit the file's contents
In trunk, the defaults for various services were added as separate files.  In these files, a 
stronger worded phrase was used to disuade site-admins from editing the file; however, the 
dcache.properties file was not updated to include this stronger phrase.

This patch updates the dcache.properties file to be in keeping with the other files in the same 
directory.

Targets: 1.9.7, trunk

Acked-by: Gerd
Patch: none
Target: Trunk
Target: 1.9.7