SrmSpaceManager: Clean File's PnfsId when TransferFinished with a failure status is returned by a pool
This is to address the ticket http://rt.dcache.org/Ticket/Display.html?id=5560

If the file reservation in a space token state needs to be change
from TRANSFERRING state back to RESERVED state due to the pool error, SpaceManager used to leave this file
reservation marked with the PnfsId which should not be associated with this file anymore.
Subsequent d deletion of the entry from the
namespace leads to this file being marked as DELETED. 

This patch corrects this behavior by removing the pnfs id from the File record when the state is changed from Transferring to Reserved. 

Another minor change is the check that AuthRecord is not null before trying to perform implicit space reservation.

Signed-Off by: Dmitry
http://rb.dcache.org/r/1726/