tests: exclude failing PinManagerTest
The PinManagerTest class provides some tests for the pin-manager; however, 
since the pin-manager is strongly coupled with PostGreSQL, the tests will only 
work if PostGreSQL is available and configured correctly.  This makes the 
PinManagerTest class fragile and it has already been disabled (by default) in 
the 1.9.6-, 1.9.7- and Trunk branches.

This patch brings 1.9.5 branch in-line with the other branches in disabling the 
PinManagerTest.  The patch is intended for 1.9.5 only.

Target: 1.9.5
Acked-by: Tigran
Patch: http://rb.dcache.org/r/1767/