Install-script: Point user to correct default directory for SSH keys
If no SSH host or server keys were found, the script was instructing
the user to call ssh-keygen with ${DCACHE_HOME} as the output directory,
which is in most cases not the directory in which the admin door actually
looks for the keys. As a simple interim solution, the directory in the
output has been corrected to dCache's default key directory, along with
a note stating that the message may be incorrectly displayed if the
keyBase property has been changed.

Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/1756/
Target: trunk
Target: 1.9.7
Target: 1.9.6
Target: 1.9.5