tests: add slf4j jar files to junit class-path
A recent patch http://rb.dcache.org/r/1751 introduced (r.13864 for 1.9.6 and
r.13867 for 1.9.5) slf4j dependencies into the JUnit tests.  Since the slf4j
jar files are not in the junit classpath for 1.9.5 and 1.9.6 branches, this
patch broke the unit tests.

This patch adds the slf4j log files to the JUnit classpath.

The patch is for 1.9.5 and 1.9.6 only; 1.9.7 and trunk already have slf4j jar
files in the junit classpath.

Target: 1.9.5
Target: 1.9.6
Patch: http://rb.dcache.org/r/1766/
Acked-by: Tirgan