web-admin: add config option to run the PoolInfoObserver cell
RT#5567 [http://rt.dcache.org/Ticket/Display.html?id=5567] requests that we start the 
PoolInfoObserver cell within the httpdDomain.  This is because the web-interface does not 
reflect changes to pools by default.  Without a PoolInfoObserver cell, the only way to 
refresh the web-interface's pool information is to restart the http domain.  Restarting 
httpdDomain is undesirable since this disrupts the info-provider, with a knock-on effect 
in the site's published information.

The proposed patch does this; however, it also allows a site to disable the feature by 
configuring the httpdDomain so it doesn't start the cell.  If the site-admin specifies a 
httpdEnablePoolCollector property in their httpdSetup file (usually a sym-link to 
dCacheSetup file) with any value other than 'yes' then the PoolInfoObserver cell will not 
be started.

Release notes: add support for starting the PoolCollector cell and start it by default.  
The PoolCollector will update the web-page with new pool information as pools change 
their configuration.  Support for this service existed in earlier releases but enabling 
it required editing the httpd.batch file.  Setting the httpdEnablePoolCollector property 
to any value other than 'yes' will disable starting the PoolCollector cell.

This patch is for 1.9.5.

Patch: http://rb.dcache.org/r/1747/
Target: 1.9.5
Acked-by: Tigran