webdav: Add "fake" lock support
Mac OS X requires lock support to enable upload to a WebDAV share. For
this reason Mac OS X clients cannot write to dCache 1.9.7 WebDAV doors.

Adding a full lock implementation is not feasible with the current
version of Milton, as there are simply too many issues regarding
authentication.

This patch adds a fake lock implementation. When requested to lock
something, a LockToken is returned, however the lock is always flagged
with a scope of NONE, a depth of ZERO and a lifetime of 0 seconds. No
attempt is made to actually lock anything at all and the server is not
promissing to do so as the locks returned have a lifetime of 0 seconds.

This seems to be enough to allow Mac OS X to upload to a dCache WebDAV
door. Once lock support in Milton matures, we will evaluate what is
required to provide a full lock implementation.

Although the patch is enough to allow Mac OS X to upload individual
files, attempts to copy a complete folder with sub-folders and files
fail. The cause of the failure is unknown, but the problem will be
investigated.

Upgraded Milton to a snapshot version of 1.5.3.

Target: trunk
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/1762/
Affects: service(webdav)