deploy: Modify migration script to generate waitForFiles configuration
To maintain the semantics of old pools blocking until the setup file is
available, this patch modifies the migration script such that a
waitForFiles setting is generated for old pools.

Notice that waitForFiles was not known prior to 1.9.7 and will not
appear in the old .poollist files.

Target: trunk
Target: 1.9.8
Target: 1.9.7
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/1790/