LoginBrokerHandler: Cache cell name and domain name
getCellName() and getDomainName() are relatively expensive calls as they
internally query the cell info. The cell info internally embeds the
output of getInfo() which in some cells are expensive to compute.
Therefore this patch caches the cell and domain name in
LoginBrokerHandler.

Eventually the mechanism for obtaining cell name and domain name should
be improved. This patch however solves the problem with a minimum of
changes and is submitted for inclusion in stable branches.

Target: trunk
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/1806/