Xrootd-door: Add support for the kXR_rm protocol message
Support for the kXR_rm protocol message has been added to the door.
Processing of that message results in deletion of the specified
regular file, if it exists. Otherwise, the door responds with an
error.

Token based authorization has been updated to parse delete-access
authorization and verify it in case of a kXR_rm call on a file.

Acked-by: Gerd Behrmann <behrmann@ndgf.org>
Patch: 1795
Target: trunk