FTP: Add LoginStrategy and update FTP doors to use it
LoginStrategy is an interface for performing login and related
operations. At the moment it only contains the login method. map and
reverse will be added later.

Two implementations of LoginStrategy are provided: GplazmaLoginStrategy
wraps the AuthzQueryHelper gPlazma cell client stub.
KauthFileLoginStrategy wraps the KAuthFile class. 

There is no LoginStrategy for using gPlazma as a module. This is so 
because we agreed to simply instantiate the gPlazma cell in all domains 
with doors configured to use the gPlazma module. Jan is working on a 
patch to update the batch files to achieve this.

Three login attributes are provided: RootDirectory, HomeDirectory and
ReadOnly.

The toString() method of our principal classes have been changed to use
the simple class name rather than the full class name. This is what is
used by the principal classes from Sun and makes log entries using the
principals easier on the eye.

The three FTP doors are updated to use the LoginStrategy. The
LoginStrategy to use is currently hardcoded in the doors to emulate the
old behaviour.

The patch destroys the "iteration over groups" behaviour of the FTP
doors. The code to perform the iteration still exists (to keep the patch
small), but it is deactivated. It will be removed in another patch.

Release notes: Only functional change I have found from this patch is
how we treat anonymous users as defined in KAuthFile for weak FTP: They
are now always represented as the NOBODY user internally. This means
they will be limited to world permissions.

Target: trunk