dcap: Ensure that gPlazma is called before responding to stat requests
In the DCAP door most requests involve a user mapping via gPlazma. This
mapping is essential for authorization. In particular when using
PnfsManager for permission checking, the mapping must be performed to
generate the Subject which is sent to PnfsManager.

For some reason the stat and lstat calls did not contain the callout to
gPlazma. This patch adds the necessary call.

Target: trunk
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/1834/