dcap: Move IoHandler specific fields to IoHandler
The access mask and isUrl fields were only used by IoHandler, yet they
were defined in PnfsSessionHandler. As a consequence it was not possible
to implement the correct semantics for _authorizationRequired being
false together with PnfsManager based permission checks: When being
false, the permission check for IoHandler should be bypassed.

This patch moves the handling of those two fields to IoHandler. This
simplifies PnfsSessionHandler and the cost of small amount of
duplication in IoHandler. The revard is that we can define a custom
PnfsHandler that does not send Subject to PnfsManager in exactly those
situations in which authorization should be bypassed.

Target: trunk
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/1853/