dcap: Use ConcurrentHashMap for sessions
The DCAP door keeps SessionHandlers in a HashMap. Unfortunately
synchronization on access to this HashMap is inconsistent and sometimes
missing.

This patch replaces the HashMap with a ConcurrentHashMap. Thus access to
the HashMap does no longer need to be synchronized.

The insertion into the map is moved to the SessionHandler. Thus
insertion and removal are now symmetric in the sense that both are
performed by the SessionHandler. As a consequence there is no need for a
message lock, as each SessionHandler will insert itself into the map
before the session will send its first message.

Finally, some duplicated code is eliminated, as check for duplicate
session IDs is now handled once in the SessionHandler.

Target: trunk
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/1857/