dcap: Introduce two step initialization for SessionHandlers
PnfsSessionHandler sends a PnfsGetFileAttributes message to PnfsManager
as part of its constructor. This poses a challenge as subclasses of
PnfsSessionHandler will not have finished their constructor and thus not
completely initialized all their fields.

In the past the dcap door used a global lock to avoid delivering the
result of the PnfsGetFileAttributes message before the constructor had
returned. That on the other hand posed the problem that the lock would
be held for a relatively long time for those handlers that performed
blocking operations in the constructor.

This patch restructures the initialization of SessionHandlers. The
constructor is responsible for parsing arguments, whereas other
operations, such as sending a PnfsGetFileAttributes message to
PnfsManager, are performed in a second initialization step triggered by
a call to the start method.

With this patch we can allow SessionHandlers to complete initialization
before any of the message handlers are called, while still avoiding
holding a global lock.

The patch also restructures some code to ensure that removeUs() is
called in case of failures.

Target: trunk
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/1864/