dcap: Use LoginStrategy
This patch updated the DCAP door to use a LoginStrategy for login
operations.

While writing this patch, I found several bugs and inconsistencies in
how the DCAP door deals with the various pieces it collects to do the
login. Rather than trying to preserve behaviour I believe is wrong, I
changed the behaviour to what I believe was the intended semantics. I
describe this in the following steps.

DCapDoorInterpreterV3 is suplied with a Subject generated by the DCAP
tunnel. This I called the authenticated Subject to distinguish it from a
Subject that was logged in.

The first DCAP command send by the client is hello. The hello command
may optionally have -uid, -gid and -pid options. -gid was not
implemented in the old DCAP door, but I have seen the dccp client
provide it and hence I implemented it. These UIDs and GIDs are never
used for authorization! Depending on the configuration of the door, they
may be used as owners for newly created name space entries.


Subsequent commands may have -uid and -gid options, too. If they do,
then those will always take precedence to the UID and GID that may have
been provided with the hello command. However the effect of the -uid and
-gid options for other commands is limited to the command for which the
options were given.

Subsequent commands may have a -role option. There was some confusion
about whether the value was an FQAN or not. I consulted old dCache 1.8
code and the value of the -role option IS NOT an FQAN. Instead, this is
a mapped user name to be used in case the credentials provided by the
user (such as Kerberos Principal or DN) map to more than one account in
gPlazma.

As the -role option may be different for each command, the login
operation (which performs mapping to UID and GID) have to be potentially
repeated for each command. In the old behaviour only the role option of
the first command that triggered a login operation would be honored, but
since the -role option can be specified with any command, I think it
makes sense to perform the login per command.

As the -role option is likely going to be the same for most commands for
a DCAP door session, a CachingLoginStrategy is used to avoid login in
the same Subject over and over.


The authorization=strong setting is used to reject any Subject that
cannot be mapped to a UID and GID. This is now enforced by using
UnionLoginStrategy, which has an option for what to do in case a UID/GID
cannot be found.

The login operation is performed as part of SessionHandler startup in
the start method. Thus the pattern of calling the old getUserMetaData
method in all com_ methods that created a SessionHandler is no longer
needed.

As all non-trivial DCAP commands cause the creation of a SessionHandler,
the creation of PnfsHandler is moved to the PnfsSessionHandler. Thus
each SessionHandler has its own PnfsHandler, which allows us to use the
Subject that resulted from the login call for that SessionHandler.

As the read only status of a session is determined not only by the door
configuration, but also by the login, the readonly check has been moved
into the session handlers. There is potential for factoring these checks
out into a common base class for non-read commands, but I propose to do
that in a subsequent patch to keep the patch size down.

Target: trunk
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/1867/