Xrootd-door: Add support for the kXR_mv protocol message
kXR_mv moves a file from one location in the filesystem to another.
In the xrootd-door this translates to a rename operation in the
namespace. getPath() was removed from AuthorizableRequestMessage, because
kXR_mv consists of two paths and if methods are not sensible for all
subclasses, inheritance gets ugly.

Acked-by: Gerd Behrmann <behrmann@ndgf.org>
Patch: 1875
Target: trunk