gplazma: plugin loader interface
This patch defines the interface for the plugin loader.  To create plugin objects, 
one uses the interface by:

1. calling PluginLoader.init method; this needs to be done once for each loader 
   instance (envisioned gPlazma usage has a single instance).

2. Obtain the plugin objects using the PluginLoader.newPluginByName method; for 
   example:

       GPlazmaPlugin plugin = loader.newPluginByName( pluginName);

There are two variants of the newPluginByName: one (illustrated above) only takes the 
name of the plugin to instantiate; the other also takes an array of String objects 
that are parsed to the plugin as arguments.  How the arguments are passed to the 
plugin is implementation-specific.

Also included is a demo / test implementation of PluginLoader, called 
StaticClassPluginLoader.  This class is instantiated with a list of plugins, each of 
which may be created using the newPluginByName method; the plugin name is just the 
simple class name.  Each plugin must have a constructor that accepts an array of 
Strings.  The no plugin-arguments newPluginByName method calls the constructor with 
an empty array.

Also included are some unit tests to check correct behaviour.

Target: trunk
Acked-by: Thomas
Acked-by: Timur