info: fix XML mark-up of attribute values
XML spec says that certain characters are not allowed; these are '&' '"' '<' and "'".  
This is true for both attribute values and plain-text.

The current XML serialiser has separate mark-up methods, on the mistaken belief that 
characters like '<' and '&' don't need to be marked-up in attribute values.

This patch fixes this mistake and marks up occurrences of '&', '"', '<', '>' and "'" in 
either plain-text or in attribute values.

The patch also updates the available unit tests for the XML serialisation.  An extra 
domain has been added (called: "bothersome"&annoying<Cells>'sDomain) with a single cell 
(called: "Foo & Bar" <foobar@example.org>'s cell).  The RELAX-NG has been updated to 
test that this domain is available in the serialised output.

Patch: http://rb.dcache.org/r/1897/
Acked-by: Gerd
Target: trunk