unit-tests: alter double asserts so they don't use deprecated form
The JUnit assertEquals() method may take two double-precision floating-point numbers and check that 
they are equal.  The deprecated method signature is:

    assertEquals( String message, double expected, double actual)

This is deprecated (and will fail tests with newer JUnit jar files) because there is no indication of 
the required accuracy of the agreement.  Except for some very specialise cases, floating-point 
arithmetic on computers is never "precise".  When testing that the desired value is achieved, one 
must say how far away the result can be before it is considered wrong.

To support this, extra assertEquals methods were introduced with the following signature:

    assertEquals( String message, double expected, double actual, double delta)

This patch adds the missing "delta" values.  Given the asserted values are constants, the arithmetic 
is simple assignments; therefore, the result is expected to agree exactly (delta=0)

Patch: http://rb.dcache.org/r/1905/
Acked-by: Thomas
Target: trunk