unit-tests: stop unit-test from being so noisy
There's a unit-test (HsmRestoreTest) that produces an awful lot of output.  As a result, looking at 
Hudson's logs is a real pain.  It turns out that this is because the unit test switches the root 
logger's threshold level to DEBUG.

The unit tests don't use the generated logging output, so setting the root logger threshold to DEBUG 
seems to be debugging code mistakenly committed.

This patch removes the code setting the root logger's threshold to DEBUG level.

Patch: http://rb.dcache.org/r/1908/
Target: trunk
Acked-by: Tigran
Acked-by: Jan