AbstractCell: Add MessageProcessingMonitor
MessageProcessingMonitor is a component used by AbstractCell to
instrument message processing with counters and time gauges.
MessageProcessingMonitor exposes a number of cell commands to query the
counters and gauges.

With this patch, any cell inheriting from AbstractCell and using the
message dispatch framework of AbstractCell will automatically get
monitored. It is however a requirement that cells use the DelayedReply
class for asynchronous message processing. Otherwise there is no way for
AbstractCell to inspect the reply.

Target: trunk
Acked-by: Jan Schaefer <jan.schaefer@desy.de>
Patch: http://rb.dcache.org/r/1878/